requirements.txt
captcha==0.1.1
Django==1.6.9
django-classy-tags==0.6.1
django-cms==3.0.12
django-mptt==0.6.1
django-sekizai==0.8.1
django-tinymce==1.5.3
Django-tinymce-filebrowser==0.2.1
djangocms-admin-style==0.2.5
djangocms-text-ckeditor==2.4.3
easy-thumbnails==2.2
html5lib==0.999
MySQL-python==1.2.3
Pillow==2.8.1
six==1.9.0
sorl-thumbnail==12.2
South==1.0.2
